+++ changelog.md
@@ -20,6 +20,7 @@ Codion Change Log
 - ComponentBuilder.name() added.
 - TableConditionPanel.get() and get(identifier) renamed panels() and panel() respectively.
 - FilterTable.createSingleSelectionModeControl() renamed createSingleSelectionControl().
+- TextComponentBuilder.initialCaretPosition() renamed caretPosition(), InitialCaretPosition renamed CaretPosition.
 ### is.codion.swing.framework.ui
 - EntityEditComponentPanel, create component methods now assign a name to the component, using attribute.toString().
 - EntityTablePanel.ControlKeys.MOVE_SELECTION_DOWN and MOVE_SELECTION_UP renamed INCREMENT_SELECTION and DECREMENT_SELECTION respectively, related renaming.